Junit4 test
