+++ .idea/libraries/junit4.xml
@@ -1,0 +1,11 @@
+<component name="libraryTable">
+  <library name="junit4" type="repository">
+    <properties maven-id="junit:junit:4.13.1" />
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/junit/junit/4.13.1/junit-4.13.1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
