+++ .idea/modules.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/OSSDP-Lab2.iml" filepath="$PROJECT_DIR$/.idea/OSSDP-Lab2.iml" />
+      <module fileurl="file://$PROJECT_DIR$/s20.iml" filepath="$PROJECT_DIR$/s20.iml" />
+    </modules>
+  </component>
+</project>
