+++ src/SolutionTest.java
@@ -1,0 +1,4 @@
+import static org.junit.Assert.*;
+public class SolutionTest {
+  
+}
